% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_namespace.R
\name{get_namespace}
\alias{get_namespace}
\title{Get Namespace}
\usage{
get_namespace(pkg)
}
\arguments{
\item{pkg}{string title}
}
\description{
Given a package title, returns the entire namespace of that package
}
