% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_namespace.R
\name{get_namespaces}
\alias{get_namespaces}
\title{Get Namespaces}
\usage{
get_namespaces(
  pkgs = c("base", "utils", "stats", "ggplot2", "dplyr", "tibble", "readr", "roxygen2",
    "devtools"),
  tofile = "data/r_namespaces.rda"
)
}
\arguments{
\item{pkgs}{vector of packages}

\item{tofile}{target file for namespaces vector output}
}
\description{
Gets the namespace for a set of packages
}
