% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purler.R
\name{id_gen}
\alias{id_gen}
\title{Generate Random IDs}
\usage{
id_gen(n = 1)
}
\arguments{
\item{n}{Number of ids to generate.}
}
\description{
Generates a given number of random, but consistent, id values. Will always
have exactly 10 characters - 5 random letters, 4 random numbers, and 1 random
trailing letter.
}
