% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purler.R
\name{purl_dir}
\alias{purl_dir}
\title{Purl Directory}
\usage{
purl_dir(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Directory with .Rmd files}

\item{output_dir}{Target directory for .R outputs}
}
\description{
Purls R code out from all .Rmd files in a given directory
}
