% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purler.R
\name{purl_filepaths}
\alias{purl_filepaths}
\title{Purl Filepaths}
\usage{
purl_filepaths(filepaths, output_dir)
}
\arguments{
\item{filepaths}{vector of filepaths to .Rmd files}

\item{output_dir}{directory to output .R code to}
}
\description{
Purls R code out from RMDs given a vector of filepaths
}
