% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode2txt.R
\name{reformat_text_R}
\alias{reformat_text_R}
\title{Reformat Text}
\usage{
reformat_text_R(path, replacements)
}
\arguments{
\item{path}{filepath to input text data (like a .R file)}

\item{replacements}{named vector, names the targets for replacements and
the values the replacement values.}
}
\description{
Function that takes a file of text and makes replacements for code 
tokenization
}
